+++ NetFabric.Hyperlinq.Benchmarks/Count.Predicate.Benchmarks.cs
@@ -52,7 +52,7 @@ namespace NetFabric.Hyperlinq.Benchmarks
 
         [BenchmarkCategory("Range")]
         [Benchmark]
-        public int Hyperlinq_Range() =>
+        public long Hyperlinq_Range() =>
             hyperlinqRange.Count((_, __) => true);
 
         [BenchmarkCategory("Queue")]
@@ -78,6 +78,6 @@ namespace NetFabric.Hyperlinq.Benchmarks
         [BenchmarkCategory("Enumerable_Value")]
         [Benchmark]
         public int Hyperlinq_Enumerable_Value() => 
-            enumerableValue.Count<TestEnumerable.Enumerable, TestEnumerable.Enumerable.Enumerator, int>((_, __) => true);
+            enumerableValue.Count<TestEnumerable.Enumerable, TestEnumerable.Enumerable.Enumerator, long>((_, __) => true);
     }
 }
