+++ NetFabric.Hyperlinq.Benchmarks/ToArray.Benchmarks.cs
@@ -31,17 +31,17 @@ namespace NetFabric.Hyperlinq.Benchmarks
 
         [BenchmarkCategory("Enumerable_Reference")]
         [Benchmark(Baseline = true)]
-        public int[] Linq_Enumerable_Reference() => 
+        public long[] Linq_Enumerable_Reference() => 
             System.Linq.Enumerable.ToArray(enumerableReference);
 
         [BenchmarkCategory("Enumerable_Value")]
         [Benchmark(Baseline = true)]
-        public int[] Linq_Enumerable_Value() => 
+        public long[] Linq_Enumerable_Value() => 
             System.Linq.Enumerable.ToArray(enumerableValue);
 
         [BenchmarkCategory("Range")]
         [Benchmark]
-        public int[] Hyperlinq_Range() =>
+        public long[] Hyperlinq_Range() =>
             hyperlinqRange.ToArray();
 
         [BenchmarkCategory("Queue")]
@@ -61,12 +61,12 @@ namespace NetFabric.Hyperlinq.Benchmarks
 
         [BenchmarkCategory("Enumerable_Reference")]
         [Benchmark]
-        public int[] Hyperlinq_Enumerable_Reference() => 
+        public long[] Hyperlinq_Enumerable_Reference() => 
             enumerableReference.ToArray();
 
         [BenchmarkCategory("Enumerable_Value")]
         [Benchmark]
-        public int[] Hyperlinq_Enumerable_Value() => 
+        public long[] Hyperlinq_Enumerable_Value() => 
             enumerableValue.ToArray();
     }
 }
